pyproject.toml
[tool.poetry]
name = "fastapi-socketio"
version = "0.10.0"
description = "Easly integrate socket.io with your FastAPI app."
license = "Apache License, Version 2.0"
authors = ["Srdjan Stankovic <stankovic.srdjo@gmail.com>"]
maintainers = ["Sumin Son <clvswft03@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/pyropy/fastapi-socketio"
repository = "https://github.com/todaypp/fastapi-socketio"
keywords = ["python", "websocket", "socketio", "fastapi", "fastapi-socketio"]
classifiers = ["Framework :: FastAPI", "Environment :: Web Environment"]
packages = [{ include = "fastapi_socketio" }]

[tool.poetry.scripts]
example-app = "examples.app:main"
example-client-threads-latency = "examples.client.threads.latency_client:main"
example-server-asgi-latency = "examples.server.asgi.latency:main"
example-server-fastapi-latency = "examples.server.fastapi.latency:main"

[tool.poetry.dependencies]
python = "~3.10"

fastapi = "0.85.0"
netifaces = "0.10.9"
pydantic = "1.10.2"
python-engineio = "3.14.2"
python-socketio = "4.6.1"
requests = "2.28.1"
starlette = "0.20.4"
uvicorn = "0.18.3"

[tool.poetry.group.dev.dependencies]
black = { version = "22.8.0", allow-prereleases = true }

[tool.poetry.group.test.dependencies]
pytest = "7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = """
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
"""
